Update: refactor folder
